+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - StoryConfig: caution, sites info
 - StoryCommand: motif, foreshadow, payoff
+- StoryConfig: copy, theme and more
 ### Changed
 - Compiler: conv to plot (re-format)
 ### Fixed